+++ scrooge-generator-tests/src/test/resources/gold_file_output_java/com/twitter/scrooge/test/gold/thriftjava/Recursive.java
@@ -276,7 +276,7 @@ public class Recursive implements TBase<Recursive, Recursive._Fields>, java.io.S
   public Object getFieldValue(_Fields field) {
     switch (field) {
     case ID:
-      return new Long(getId());
+      return Long.valueOf(getId());
     case REC_REQUEST:
       return getRecRequest();
     }